Add files via upload
I Create a function that reads a CSV file. It accepts the name of the file (str), returns a list of strings (list). Also create a function that writes data to a file. It accepts the name of the file (str), a list of strings (list) that need to be written to the file. Does not return anything. Now, using the created functions, first create a file (3 lines are enough), then read the same file, write the lines in a variable, then add two lines to the variable, and then write your changes in the second file.

II Create a function that returns the squares of all numbers in the range 0 - 100000.
It should work and not put your computer down.